some fixes

+++ VK.Bot.ConsoleClient/Options/AdderOptions.cs
@@ -5,9 +5,9 @@ namespace VK.Bot.ConsoleClient.Options
     public class AdderOptions
     {
         [Value(0, Required = true)]
-        public int ValueOne { get; set; }
+        public double ValueOne { get; set; }
 
         [Value(1, Required = true)]
-        public int ValueTwo { get; set; }
+        public double ValueTwo { get; set; }
     }
 }
